LITE-28473 Columns can have different types in manual transformation. Unification of output column input
