+++ connect_transformations/static/transformations/currency_conversion.html
@@ -1,1 +1,1 @@
-<!doctype html><html><head><meta charset="utf-8"/><title>Transformations/Currency conversion</title><script defer="defer" src="../vendors.b7829ba6a3fa12b6e5cb.js"></script><script defer="defer" src="../transformations/currency_conversion.681193f69422b0013fba.js"></script><link href="../transformations/currency_conversion.a15ed0d89abdd4879a88.css" rel="stylesheet"></head><body><div id="loader"></div><div id="app"><div class="main-container" id="content"></div><div class="button-container"><button id="add" class="button">ADD COLUMN</button></div></div></body></html>
+<!doctype html><html><head><meta charset="utf-8"/><title>Transformations/Currency conversion</title><script defer="defer" src="../vendors.b7829ba6a3fa12b6e5cb.js"></script><script defer="defer" src="../transformations/currency_conversion.2aac1b5a43dbb030c0a3.js"></script><link href="../transformations/currency_conversion.8fb9bfd4093578f7ab2e.css" rel="stylesheet"></head><body><div id="loader"></div><div id="app"><div class="main-container" id="content"></div><div class="button-container"><button id="add" class="button">ADD COLUMN</button></div></div></body></html>
