+++ connect_transformations/static/transformations/formula.html
@@ -1,1 +1,1 @@
-<!doctype html><html lang="en"><head><meta charset="utf-8"/><title>Transformations/Formula</title><script defer="defer" src="../vendors.b7829ba6a3fa12b6e5cb.js"></script><script defer="defer" src="../transformations/formula.6f18005ff74300bec934.js"></script><link href="../transformations/formula.c92c87f88e57181ad946.css" rel="stylesheet"></head><body><div id="loader"></div><div id="app"><div class="main-container" id="content"></div><div class="button-container"><button id="add" class="button">+ ADD</button></div></div></body></html>
+<!doctype html><html lang="en"><head><meta charset="utf-8"/><title>Transformations/Formula</title><script defer="defer" src="../vendors.b7829ba6a3fa12b6e5cb.js"></script><script defer="defer" src="../transformations/formula.781b67ff03a92d4e382d.js"></script><link href="../transformations/formula.96a5053f4ed12985abe8.css" rel="stylesheet"></head><body><div id="loader"></div><div id="app"><div class="main-container" id="content"></div><div class="button-container"><button id="add" class="button">ADD</button></div></div></body></html>
