+++ connect_transformations/static/transformations/lookup_subscription.html
@@ -1,1 +1,1 @@
-<!doctype html><html><head><meta charset="utf-8"/><title>Transformations/Lookup subscription</title><script defer="defer" src="../vendors.b7829ba6a3fa12b6e5cb.js"></script><script defer="defer" src="../transformations/lookup_subscription.37ee54e243c32076b1a5.js"></script><link href="../transformations/lookup_subscription.b3273c9e08d074413c4d.css" rel="stylesheet"></head><body><div id="loader"></div><div id="app"><div class="main-container lookup" id="content"><div class="subtitle">Supported search criteria include Subscription External ID, ID and Parameter values.</div><div class="input-group"><label class="label" for="criteria">Criterion</label> <select class="input" id="criteria"></select></div><div class="input-group" id="param_name_group"><label class="label" for="parameter">Parameter Name</label> <select class="input" id="parameter"></select></div><div class="input-group"><label class="label" for="column">Input Column</label> <select class="input" id="column"></select></div><div class="input-group"><label class="label" for="prefix">Prefix</label> <input class="input" id="prefix" placeholder="Prefix"/></div><div class="input-group"><p class="radio-buttons">Choose what to do if a subscription is not found:</p><div class="radio-buttons input"><input type="radio" id="not_found_leave_empty" name="if_not_found" value="leave_empty"/> <label class="label" for="not_found_leave_empty">Leave the row empty</label> <input type="radio" id="not_found_fail" name="if_not_found" value="fail"/> <label class="label" for="not_found_fail">Transformation will fail</label></div></div><div class="input-group"><p class="radio-buttons">Choose what to do if multiple subscriptions are found:</p><div class="radio-buttons input radio-buttons-vertical"><div><input type="radio" id="multiple_use_most_actual" name="if_multiple" value="use_most_actual"/> <label class="label" for="multiple_use_most_actual">Use the most recent subscription</label></div><div><input type="radio" id="multiple_leave_empty" name="if_multiple" value="leave_empty"/> <label class="label" for="multiple_leave_empty">Leave the row empty</label></div><div><input type="radio" id="multiple_fail" name="if_multiple" value="fail"/> <label class="label" for="multiple_fail">Transformation will fail</label></div></div></div></div></div></body></html>
+<!doctype html><html><head><meta charset="utf-8"/><title>Transformations/Lookup subscription</title><script defer="defer" src="../vendors.b7829ba6a3fa12b6e5cb.js"></script><script defer="defer" src="../transformations/lookup_subscription.343caa83ec9312471a6a.js"></script><link href="../transformations/lookup_subscription.10c101d9f62c0cae5e02.css" rel="stylesheet"></head><body><div id="loader"></div><div id="app"><div class="main-container lookup" id="content"><div class="subtitle">Supported search criteria include Subscription External ID, ID and Parameter values.</div><div class="input-group"><label class="label" for="criteria">Criterion</label> <select class="input" id="criteria"></select></div><div class="input-group" id="param_name_group"><label class="label" for="parameter">Parameter Name</label> <select class="input" id="parameter"></select></div><div class="input-group"><label class="label" for="column">Input Column</label> <select class="input" id="column"></select></div><div class="input-group"><label class="label" for="prefix">Prefix</label> <input class="input" id="prefix" placeholder="Prefix"/></div><div class="input-group"><p class="radio-buttons">Choose what to do if a subscription is not found:</p><div class="radio-buttons input"><input type="radio" id="not_found_leave_empty" name="if_not_found" value="leave_empty"/> <label class="label" for="not_found_leave_empty">Leave the row empty</label> <input type="radio" id="not_found_fail" name="if_not_found" value="fail"/> <label class="label" for="not_found_fail">Transformation will fail</label></div></div><div class="input-group"><p class="radio-buttons">Choose what to do if multiple subscriptions are found:</p><div class="radio-buttons input radio-buttons-vertical"><div><input type="radio" id="multiple_use_most_actual" name="if_multiple" value="use_most_actual"/> <label class="label" for="multiple_use_most_actual">Use the most recent subscription</label></div><div><input type="radio" id="multiple_leave_empty" name="if_multiple" value="leave_empty"/> <label class="label" for="multiple_leave_empty">Leave the row empty</label></div><div><input type="radio" id="multiple_fail" name="if_multiple" value="fail"/> <label class="label" for="multiple_fail">Transformation will fail</label></div></div></div></div></div></body></html>
