+++ connect_transformations/static/transformations/manual.html
@@ -1,1 +1,1 @@
-<!doctype html><html><head><meta charset="utf-8"/><title>Transformations/Manual</title><script defer="defer" src="../vendors.b7829ba6a3fa12b6e5cb.js"></script><script defer="defer" src="../transformations/manual.d9528ab0180232435b13.js"></script><link href="../transformations/manual.ce8032b639007f4c0ddb.css" rel="stylesheet"></head><body><div id="loader"></div><div id="app" class="hidden"><div class="main-container" id="content"><div id="description"><h5 id="description-label" class="card-header">Instructions:</h5><textarea id="description-text" placeholder="Text of instructions" rows="6"></textarea><div class="description-text-details"><div id="description-text-hint" class="message error--text text-hidden">Input must be at most 2000 characters</div><div id="description-text-counter" class="counter">0 / 2000</div></div></div><div id="edit-input"><h5 class="card-header">Input columns:</h5><table id="edit-input-columns-table"><thead><tr><th>ID</th><th>NAME</th><th>TYPE</th><th>DESCRIPTION</th><th></th></tr></thead><tbody id="edit-input-columns"></tbody></table></div><div id="settings"><h5 class="card-header">Settings:</h5><textarea id="settings-text" rows="6">{}</textarea></div><div id="output"><h5 class="card-header">Output columns:</h5><div id="output-columns"></div><div class="button-container"><button id="add" class="button">ADD COLUMN</button></div></div></div></div></body></html>
+<!doctype html><html><head><meta charset="utf-8"/><title>Transformations/Manual</title><script defer="defer" src="../vendors.b7829ba6a3fa12b6e5cb.js"></script><script defer="defer" src="../transformations/manual.314d3c570319a7ee98dd.js"></script><link href="../transformations/manual.acb3c960dc0e6a562c3e.css" rel="stylesheet"></head><body><div id="loader"></div><div id="app" class="hidden"><div class="main-container" id="content"><div id="description"><h5 id="description-label" class="card-header">Instructions:</h5><textarea id="description-text" placeholder="Text of instructions" rows="6"></textarea><div class="description-text-details"><div id="description-text-hint" class="message error--text text-hidden">Input must be at most 2000 characters</div><div id="description-text-counter" class="counter">0 / 2000</div></div></div><div id="edit-input"><h5 class="card-header">Input columns:</h5><table id="edit-input-columns-table"><thead><tr><th>ID</th><th>NAME</th><th>TYPE</th><th>DESCRIPTION</th><th></th></tr></thead><tbody id="edit-input-columns"></tbody></table></div><div id="settings"><h5 class="card-header">Settings:</h5><textarea id="settings-text" rows="6">{}</textarea></div><div id="output"><h5 class="card-header">Output columns:</h5><div id="output-columns"></div><div class="button-container"><button id="add" class="button">ADD</button></div></div></div></div></body></html>
