+++ connect_transformations/static/transformations/split_columns.html
@@ -1,1 +1,1 @@
-<!doctype html><html><head><meta charset="utf-8"/><title>Transformations/Split columns</title><script defer="defer" src="../vendors.b7829ba6a3fa12b6e5cb.js"></script><script defer="defer" src="../transformations/split_columns.79b268b6c97286981208.js"></script><link href="../transformations/split_columns.4e82789c8bb0f422cdc8.css" rel="stylesheet"></head><body><div id="loader"></div><div id="app"><div class="main-container" id="content"><div id="input-column" class="input-group"><label class="label" for="column">Input Column</label> <select class="input" id="column"></select></div><div id="regex-column" class="input-group"><label class="label" for="pattern">Regular Expression</label> <input id="pattern" placeholder="Regular expression"/></div><div id="output-column" class="input-group"><label class="label" for="output">Output Columns</label><div id="output"></div><button type="button" id="refresh" class="button">Validate</button></div></div></div></body></html>
+<!doctype html><html><head><meta charset="utf-8"/><title>Transformations/Split columns</title><script defer="defer" src="../vendors.b7829ba6a3fa12b6e5cb.js"></script><script defer="defer" src="../transformations/split_columns.07cc299307f9b2495d57.js"></script><link href="../transformations/split_columns.07553b4314d9bf83f149.css" rel="stylesheet"></head><body><div id="loader"></div><div id="app"><div class="main-container" id="content"><div id="input-column" class="input-group"><label class="label" for="column">Input Column</label> <select class="input" id="column"></select></div><div id="regex-column" class="input-group"><label class="label" for="pattern">Regular Expression</label> <input id="pattern" placeholder="Regular expression"/></div><div id="output-column" class="input-group"><label class="label" for="output">Output Columns</label><div id="output"></div><button type="button" id="refresh" class="button">Validate</button></div></div></div></body></html>
