+++ connect_transformations/static/transformations/vat_rate.html
@@ -1,1 +1,1 @@
-<!doctype html><html><head><meta charset="utf-8"/><title>Transformations/VAT rate</title><script defer="defer" src="../vendors.b7829ba6a3fa12b6e5cb.js"></script><script defer="defer" src="../transformations/vat_rate.2341283a034df19516a8.js"></script><link href="../transformations/vat_rate.28c6812c7e0529812293.css" rel="stylesheet"></head><body><div id="loader"></div><div id="app"><div class="main-container" id="content"><div class="wrapper"><div class="input-group"><label for="input-column">Input Column</label> <select name="inputColumn" id="input-column"></select></div><div class="input-group"><label for="output-column">Output Column</label> <input name="outputColumn" id="output-column"></div></div><div class="row"><p class="radio-buttons">Choose what to do when an incorrect country code or a non-EU country value is passed</p><div class="radio-buttons"><input type="radio" id="leave_empty" name="if_not_found" value="leave_empty"/> <label class="label" for="leave_empty">Leave the row empty</label> <input type="radio" id="fail" name="if_not_found" value="Fail"/> <label class="label" for="fail">Transformation will fail</label></div></div></div></div></body></html>
+<!doctype html><html><head><meta charset="utf-8"/><title>Transformations/VAT rate</title><script defer="defer" src="../vendors.b7829ba6a3fa12b6e5cb.js"></script><script defer="defer" src="../transformations/vat_rate.9ff38b1f97c76f225b02.js"></script><link href="../transformations/vat_rate.423ffa54d40daf6f2f4d.css" rel="stylesheet"></head><body><div id="loader"></div><div id="app"><div class="main-container" id="content"><div class="wrapper"><div class="input-group"><label for="input-column">Input Column</label> <select name="inputColumn" id="input-column"></select></div><div class="input-group"><label for="output-column">Output Column</label> <input name="outputColumn" id="output-column"></div></div><div class="row"><p class="radio-buttons">Choose what to do when an incorrect country code or a non-EU country value is passed</p><div class="radio-buttons"><input type="radio" id="leave_empty" name="if_not_found" value="leave_empty"/> <label class="label" for="leave_empty">Leave the row empty</label> <input type="radio" id="fail" name="if_not_found" value="Fail"/> <label class="label" for="fail">Transformation will fail</label></div></div></div></div></body></html>
