+++ ui/pages/transformations/formula.html
@@ -18,7 +18,7 @@ All rights reserved.
         <div class="main-container" id="content">
         </div>
         <div class="button-container">
-            <button id="add" class="button">+ ADD</button>
+            <button id="add" class="button">ADD</button>
         </div>
     </div>
 </body>
