+++ ui/pages/transformations/manual.html
@@ -50,7 +50,7 @@ All rights reserved.
                 <div id="output-columns">
                 </div>
                 <div class="button-container">
-                    <button id="add" class="button">ADD COLUMN</button>
+                    <button id="add" class="button">ADD</button>
                 </div>
             </div>
         </div>
