+++ ui/styles/split_column.css
@@ -9,6 +9,7 @@
     flex-direction: column;
 }
 
+#column,
 #output {
     width: 100%;
 }
@@ -19,16 +20,6 @@
     justify-content: space-evenly;
     align-items: center;
     width: 466px;
-}
-
-input {
-    width: 100%;
-    padding: 8px 10px;
-}
-
-select {
-    width: 100%;
-    padding: 8px 10px;
 }
 
 .label {
@@ -54,13 +45,6 @@ select {
 
 #refresh {
     margin: 10px 0;
-}
-
-.wrapper {
-    display: grid;
-    grid-template-columns: 1fr 1fr 1fr;
-    justify-items: center;
-    grid-gap: 5px;
 }
 
 .output-header {
